Move allocator components to clearing sub-pkg
